parley: Move GlyphRenderer trait into parley module
